+++ tests/Features/ReturnTypes/CarbonImmutable/bootstrap.php
@@ -1,0 +1,6 @@
+<?php
+
+use Carbon\CarbonImmutable;
+use Illuminate\Support\Facades\Date;
+
+Date::use(CarbonImmutable::class);