+++ tests/FeaturesTest.php
@@ -41,7 +41,7 @@ class FeaturesTest extends BaseTestCase
         $baseDir = __DIR__.DIRECTORY_SEPARATOR.'Features'.DIRECTORY_SEPARATOR;
 
         /** @var SplFileInfo $file */
-        foreach ((new Finder())->in($baseDir)->files()->name('*.php') as $file) {
+        foreach ((new Finder())->in($baseDir)->files()->name('*.php')->notName('bootstrap.php') as $file) {
             $fullPath = realpath((string) $file);
             $calls[str_replace($baseDir, '', $fullPath)] = [$fullPath];
         }
@@ -54,6 +54,11 @@ class FeaturesTest extends BaseTestCase
      */
     public function testFeatures(string $file): void
     {
+        $configFile = dirname($file).'/phpstan-tests.neon';
+        if (file_exists($configFile)) {
+            $this->configPath = $configFile;
+        }
+
         if ($this->analyze($file) === 0) {
             $this->assertTrue(true);
         }