+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Improvements
 * feat: add Storage::drive() support by @PrinsFrank in https://github.com/nunomaduro/larastan/pull/1241
 
+### Improvements
+* feat: Detect actual date class in Date facade, model attributes, now and today helpers by @bastien-phi in https://github.com/nunomaduro/larastan/pull/1142
+
 ### Fixes
 
 * fix: Change QueryBuilder::newQuery() @return from `$this` to `static`